+++ packages/broker/test/unit/plugins/websocket/WebsocketServer.test.ts
@@ -1,9 +1,11 @@
 import WebSocket from 'ws'
 import qs from 'qs'
-import StreamrClient from 'streamr-client'
+import StreamrClient, { Subscription } from 'streamr-client'
 import { waitForEvent, waitForCondition } from '@streamr/utils'
 import { WebsocketServer } from '../../../../src/plugins/websocket/WebsocketServer'
 import { PlainPayloadFormat } from '../../../../src/helpers/PayloadFormat'
+import { mock, MockProxy } from 'jest-mock-extended'
+import { merge } from '@streamr/utils'
 
 const PORT = 12398
 const MOCK_STREAM_ID = '0x1234567890123456789012345678901234567890/mock-path'
@@ -15,10 +17,14 @@ const PATH_SUBSCRIBE_MOCK_STREAM = `/streams/${encodeURIComponent(MOCK_STREAM_ID
 const REQUIRED_API_KEY = 'required-api-key'
 
 const createTestClient = (path: string, queryParams?: any): WebSocket => {
-    const queryParamsSuffix = qs.stringify({
-        apiKey: REQUIRED_API_KEY,
-        ...queryParams
-    })
+    const queryParamsSuffix = qs.stringify(
+        merge(
+            {
+                apiKey: REQUIRED_API_KEY
+            },
+            queryParams
+        )
+    )
     return new WebSocket(`ws://localhost:${PORT}${path}?${queryParamsSuffix}`)
 }
 
@@ -26,7 +32,7 @@ describe('WebsocketServer', () => {
 
     let wsServer: WebsocketServer
     let wsClient: WebSocket
-    let streamrClient: Partial<StreamrClient>
+    let streamrClient: MockProxy<StreamrClient>
 
     const assertConnectionError = async (expectedHttpStatus: number) => {
         const [ error ] = await waitForEvent(wsClient, 'error')
@@ -34,11 +40,9 @@ describe('WebsocketServer', () => {
     }
 
     beforeEach(async () => {
-        streamrClient = {
-            publish: jest.fn().mockResolvedValue(undefined),
-            subscribe: jest.fn().mockResolvedValue(undefined),
-        } as Partial<StreamrClient>
-        wsServer = new WebsocketServer(streamrClient as StreamrClient, 0, 0)
+        streamrClient = mock<StreamrClient>()
+        streamrClient.subscribe.mockResolvedValue(mock<Subscription>())
+        wsServer = new WebsocketServer(streamrClient, 0, 0)
         await wsServer.start(PORT, new PlainPayloadFormat(), {
             keys: [REQUIRED_API_KEY]
         })
@@ -62,15 +66,15 @@ describe('WebsocketServer', () => {
 
     describe('publish', () => {
 
-        const publish = async (queryParams?: any) => {
+        const connectAndPublish = async (queryParams?: any) => {
             wsClient = createTestClient(PATH_PUBLISH_MOCK_STREAM, queryParams)
             await waitForEvent(wsClient, 'open')
             wsClient.send(JSON.stringify(MOCK_MESSAGE))
-            await waitForCondition(() => ((streamrClient.publish as any).mock.calls.length === 1))
+            await waitForCondition(() => (streamrClient.publish.mock.calls.length === 1))
         }
 
         it('without parameters', async () => {
-            await publish()
+            await connectAndPublish()
             expect(streamrClient.publish).toBeCalledWith(
                 {
                     id: MOCK_STREAM_ID,
@@ -84,11 +88,11 @@ describe('WebsocketServer', () => {
         })
 
         it('valid partition', async () => {
-            await publish({ partition: 123 })
+            await connectAndPublish({ partition: 50 })
             expect(streamrClient.publish).toBeCalledWith(
                 {
                     id: MOCK_STREAM_ID,
-                    partition: 123
+                    partition: 50
                 }, 
                 MOCK_MESSAGE,
                 {
@@ -98,7 +102,7 @@ describe('WebsocketServer', () => {
         })
 
         it('valid partitionKey', async () => {
-            await publish({ partitionKey: 'mock-key' })
+            await connectAndPublish({ partitionKey: 'mock-key' })
             expect(streamrClient.publish).toBeCalledWith(
                 {
                     id: MOCK_STREAM_ID,
@@ -113,7 +117,7 @@ describe('WebsocketServer', () => {
         })
 
         it('valid partitionKeyField', async () => {
-            await publish({ partitionKeyField: 'foo' })
+            await connectAndPublish({ partitionKeyField: 'foo' })
             expect(streamrClient.publish).toBeCalledWith(
                 {
                     id: MOCK_STREAM_ID,
@@ -170,10 +174,29 @@ describe('WebsocketServer', () => {
             const payloadPromise = waitForEvent(wsClient, 'message')
             const onMessageCallback = (streamrClient.subscribe as any).mock.calls[0][1]
             onMessageCallback(MOCK_MESSAGE, {})
-            const firstPayload = (await payloadPromise)[0] as string
-            expect(JSON.parse(firstPayload)).toEqual(MOCK_MESSAGE)
+            const firstPayload = (await payloadPromise)[0]
+            expect(JSON.parse(firstPayload as string)).toEqual(MOCK_MESSAGE)
         })
 
+        it('if client#subscribe throws, passed subscriptions are cleaned', async () => {
+            const singletonSubscription = mock<Subscription>()
+            streamrClient.subscribe
+                .mockResolvedValueOnce(singletonSubscription)
+                .mockResolvedValueOnce(singletonSubscription)
+                .mockRejectedValue(new Error('bad partition'))
+            wsClient = createTestClient(PATH_SUBSCRIBE_MOCK_STREAM, { partitions: '0,2,350' })
+            await waitForEvent(wsClient, 'close')
+            expect(singletonSubscription.unsubscribe).toHaveBeenCalledTimes(2)
+        })
+
+        it('on client disconnect subscriptions are cleaned', async () => {
+            const singletonSubscription = mock<Subscription>()
+            streamrClient.subscribe.mockResolvedValue(singletonSubscription)
+            wsClient = createTestClient(PATH_SUBSCRIBE_MOCK_STREAM, { partitions: '0,2,5' })
+            await waitForEvent(wsClient, 'open')
+            wsClient.close()
+            await waitForCondition(() => singletonSubscription.unsubscribe.mock.calls.length === 3)
+        })
     })
 
 })